Iteration 1
- Add base readme and license files

- Add basic strucuture for launch file in python

- Added two packages that implement ROS2 launch policies with python and prepared a Husky robot from Ignition to use for research.

- Moved resources path and fix basic python launch file

- Managed the resources path to the workspace and fixed the .py example launch file.

- Added joint state publisher to robot

- Added gazebo to ros parameter bridge for cmd_vel

- Added resources path

- Added an empty world with a husky robot almost opperational.

- added control file

- Added more complex worlds

- Added transform for lidar and introduce TF tree standard

- Ros2 does not allow to parameterize namespaces in SDF files, which prevents manipulating tf prefixes in a multi-robot setting.
Therefore, in this commit I've added the topic mapping from gazebo to ROS and a TF naming convention to allow launching multiple robots within ros2 framework without changing extra structures.
Nodes such as the robot_state_publisher add prefixes to the entiire tree in multi-robot configuration, which might not be aligned with what sensors are publishing from gazebo.
For example, a lidar sensor associate readings with transformations that does not exist and preventing visualization in Rviz.
This may break your robot's functioning, unless you have understanding about the TF trees and how they are handled and provide a worksaround.
With this commit, one can easily handle multi-robots in ros2 with the correct transformation trees and individual topics.

- Add parameterized launch for the Husky robot

- Added a parameterized version to launch multiple robots.

- In ROS2, the Python launch file architecture is evaluated only in runtime.
Therefore, there is the need to replace parameters when launching actions.
In this regard, there is no easy way to parameterize launch files without them and
there aren't methods that allow simple text replacement.
For example, consider a situation where you need to change the namespace of a robot when launching it.
In this situation, you need to remap and add gz_bridge, if using simulation, strings that
are passed to nodes, such as the teleop. However, ros does not allow you to simple pass the received argument
the way the launch file receive it. This commit added a workaround with PythonExpressions to replace
single ocurrences by a LaunchConfiguration object.

- Fix multiple robots launch file

- This commit addressed the issue of lack of a parameterized working launch file for the husky robot.
In particullar, the parameterized launch file was fixed and now it allows to spawn multiple working husky robots.
They are ready for research and use with CMD vel and LIDAR topics being automatically mapped to ROS2 from GAZEBO IGN.
Also, the transformations are properly set considering each robot namespace.

- Fix launch files

0 In ROS2 python launch files should end with '_launch.py' to allow ros programs to see them.
In this commit I've fixed the launch files and added a setup.py to make colcon or any other
ros related procedure knowns where they are.

Renamed the package to multi-robot-simulations, for two reasons.

1 - During my interaction with several students and professors I've realised that
people are often unable to understand and use ROS and ROS2 properly. Therefore, I found better,
according to the best of my expertise, to create a separate package only with simulations that are
ready to be used for research. The intuition is to make things easier for a newbie, only exposing the necessary
topics for robotics research regarding any autonomy stack. I crafted these environments so each robot is parameterized with
its own namespace, have the correct transformation tree, and parameterized the gazebo bridge for ROS2.

2 - To separate several simulation environments with different robots and how they should be configured to work with ROS2.

- Changed the launch files names and fixed them inside the bringup for multi and single robots.

- Added the launch folder again. For some reason
it was removed from the repo.

- Fix single robot with config file name

- Fix single robot launch file

- Fix the path of a node in the single robot launch example.

- Added more robots, such as drones and smaller cars

- Removed the assets from the package folder and added more robots into the simulations.

- Removed unused SPOT model and initiated fixinig robot controller for X4 drone

- Fix the namespace replacement for the X4 drone in the SDF file

- In ROS 2, the substitution from the python launch files prevents one to easily replace
variables inside documents or strings, simply because they are evaluated during runtime.
To overcome this limitation in the Multicopter motor plugin, I've added a PythonExpression
substitution that receives the SDF string and put the action object from the namespace
received as a launch argument stored inside a LaunchConfiguration object.

- Added gz bridge for X2 and X4 drone and fix lidar pose for Husky and X2

- I've added the transformations for lidar sensor in the X2 robot and also
applied a transformation in the Husky lidar to obtain a more accurate point cloud.

- Removed extra joint since it was causing instabilities in drone controller

- Removed dummy link from Husky robot

For some reason it was useless to have this link inside the robot's sdf file in ROS 2,
because the ROS stack was not ignoring its innertia component.

- Add pose publisher to all robots

- Added the pose publisher to all robots to have
necessary information for simulation purposes. With the pose
one can implement a pose graph based map construction algorithm without the necessity
to run a SLAM algorithm. Consequently, the simulations can run much smoother and
yet have a certain degree of realism regarding how the map is being built and
shared among different robots through invidiaul readings. Furthermore,
with the real pose from Gazebo, one is able to assemble any kind of map
with annotated point clouds and work semantic mapping.

- Added transform node structure and object oriented architecture for ROS 2

- Prepared the cMakeLists to create a ROS 2 node in a separate header file.
This header is then included in a main file that is able to load all nodes.
There is still job to make it work with the pose callback.

- Added a tf publisher for the odom frame

- Added a transform publisher using the ROS2 architecture
by extending the Node class. There is still an issue, where the RViz2
is unable to find the transform from odom to base_link to show
point clouds.

- Changed odom and laser transform frames to fixed

- Changed the odom transform publisher to a
static transform publisher. This fixed the bug where RViz
is unable to find the transform link between the laser and
the odom frame.

- Fix laser transform for Husky and add Odom publisher to all robots

- Fix the transform for the Husky robot lidar relative
to the robot's frame instead of the sensor rack. Furthermore, I've also
added the odom publisher to all robots. They are now ready to be used
for research. Next steps include creating rViz configurations to each robot.

- Added config files and fix pose message from Odom publisher

- Added a rViz configuration file for each robot from
the base simulation. Furthermore, I've removed the 'seeking pose'
message that was being printed inside the TF publisher node.

- Added structure of the documentation

- Added more content for the documentation

- Published a more polished version of the statement and started
adding content to the working environment.