Add transform for lidar and introduce TF tree standard
Ros2 does not allow to parameterize namespaces in SDF files, which prevents manipulating tf prefixes in a multi-robot setting.
Therefore, in this commit I've added the topic mapping from gazebo to ROS and a TF naming convention to allow launching multiple robots within ros2 framework without changing extra structures.
Nodes such as the robot_state_publisher add prefixes to the entiire tree in multi-robot configuration, which might not be aligned with what sensors are publishing from gazebo.
For example, a lidar sensor associate readings with transformations that does not exist and preventing visualization in Rviz.
This may break your robot's functioning, unless you have understanding about the TF trees and how they are handled and provide a worksaround.
With this commit, one can easily handle multi-robots in ros2 with the correct transformation trees and individual topics.