Add parameterized launch for the Husky robot
In this commit I've added a parameterized version to launch multiple robots.
In ROS2, the Python launch file architecture is evaluated only in runtime.
Therefore, there is the need to replace parameters when launching actions.
In this regard, there is no easy way to parameterize launch files without them and
there aren't methods that allow simple text replacement.
For example, consider a situation where you need to change the namespace of a robot when launching it.
In this situation, you need to remap and add gz_bridge, if using simulation, strings that
are passed to nodes, such as the teleop. However, ros does not allow you to simple pass the received argument
the way the launch file receive it. This commit added a workaround with PythonExpressions to replace
single ocurrences by a LaunchConfiguration object.