Fix multiple robots launch file
This commit addressed the issue of lack of a parameterized working launch file for the husky robot.
In particullar, the parameterized launch file was fixed and now it allows to spawn multiple working husky robots.
They are ready for research and use with CMD vel and LIDAR topics being automatically mapped to ROS2 from GAZEBO IGN.
Also, the transformations are properly set considering each robot namespace.