Rename the project to multi-robot-simulations
I've renamed the package to multi-robot-simulations, for two reasons.

1 - During my interaction with several students and professors I've realised that
people are often unable to understand and use ROS and ROS2 properly. Therefore, I found better,
according to the best of my expertise, to create a separate package only with simulations that are
ready to be used for research. The intuition is to make things easier for a newbie, only exposing the necessary
topics for robotics research regarding any autonomy stack. I crafted these environments so each robot is parameterized with
its own namespace, have the correct transformation tree, and parameterized the gazebo bridge for ROS2.

2 - To separate several simulation environments with different robots and how they should be configured to work with ROS2.