Fix the namespace replacement for the X4 drone in the SDF file
In ROS 2, the substitution from the python launch files prevents one to easily replace
variables inside documents or strings, simply because they are evaluated during runtime.
To overcome this limitation in the Multicopter motor plugin, I've added a PythonExpression
substitution that receives the SDF string and put the action object from the namespace
received as a launch argument stored inside a LaunchConfiguration object.