Fix laser transform for Husky and add Odom publisher to all robots
In this commit I've fixed the transform for the Husky robot lidar relative
to the robot's frame instead of the sensor rack. Furthermore, I've also
added the odom publisher to all robots. They are now ready to be used
for research. Next steps include creating rViz configurations to each robot.